analysis/requirements.txt
awscli==1.6.2
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
colorama==0.2.5
cycler==0.10.0
decorator==3.4.0
docutils==0.12
functools32==3.2.3.post2
genson==0.1.0
GraphLab-Create==1.8.5
jmespath==0.5.0
jsonschema==2.5.0
matplotlib==1.5.1
mixpanel-py==3.1.1
multipledispatch==0.4.8
numpy==1.11.0
pandas==0.18.0
prettytable==0.7.2
pyasn1==0.1.9
pyparsing==2.1.1
python-dateutil==2.5.2
pytz==2016.3
requests==2.3.0
rsa==3.1.2
six==1.10.0
sseclient==0.0.8
tornado==4.1
